test: add reducer tests
